refactor(worker): give the daily report's adoption domain its own owner (Part of #1838)
Chunk 2 of the #1838 merge. Moves yesterday's install / onboarding / activation
/ engine / polish-provider / volume / geography domain out of
`daily-report/src/index.js` into `daily-report/src/adoption.js`. Queries,
populations, result shapes, degradation behaviour and rendered content are
unchanged — this moves ownership, nothing else.

`adoption.js` owns the seven adoption queries, `activeUsersSubquery`,
`tierASqlFor`, `fetchReportData` and `resolveBuckets`. `buildMessage` and its
label / percentage / date helpers deliberately stay put: presentation moves once,
in the chunk that gives it its own owner, rather than being relocated twice.

`sqlTimestamp` and `windowClause` moved to `lib/posthog.js` rather than staying
in `index.js` as first briefed. Keeping them there would have forced
`adoption.js` to import from `index.js` while `index.js` imported
`fetchReportData` back — a deliberate import cycle in a change whose whole point
is clean ownership, which the version scorecard would then have inherited. They
are generic HogQL construction with no adoption-specific population or product
judgement, so they belong beside `sqlIdList`.

Four source-guardrail tests read the worker source as TEXT and assert every
query still carries the shared production predicate — the guards that exist
because a duplicated dev-exclusion subquery caused real PostHog timeouts twice
(#1655, #1716). Moving the SQL made all four fail, since they were reading a file
that no longer holds the queries. They are REPOINTED at `adoption.js`, not
weakened: a guard that passes because it is inspecting the wrong file is worse
than one that fails.

No forwarding shims; tests and the live smoke import each domain from its owner.

Verification: node --test expected 74, actual 74, pass 74, fail 0 — the delta
from 72 is exactly the two characterization tests for the relocated SQL helpers.
10 consecutive runs, zero failures. node --check clean on all four touched files.
wrangler deploy --dry-run clean at 19.16 KiB. No index.js/adoption.js cycle.

Self-audit-grep: grep -n "from \"./index.js\"" workers/daily-report/src/adoption.js (0 hits, no cycle); grep -rn "DEFAULT_ENGINE\|DEFAULT_PROVIDER" workers/daily-report/src/index.js workers/daily-report/test (0 hits before narrowing them to module-private)
Self-audit-owner: workers/daily-report/src/adoption.js
